producer/requirements.txt
Flask
flask-cors
pika
mysql-connector-python
python-dotenv
jsonschema